Re-organize benches
